better splitting up of the components with asides
